pyproject.toml
[tool.poetry]
name = "census"
version = "0.1.0"
description = "Secure, server-rendered survey platform for medical audit and research"
authors = ["Census Maintainers <maintainers@example.com>"]
readme = "README.md"
packages = [
    { include = "census_app" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
django = "^5.1"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.9"
jinja2 = "^3.1.4"
django-cors-headers = "^4.4.0"
whitenoise = "^6.7.0"
django-axes = "^6.5.0"
django-ratelimit = "^4.1.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
cryptography = "^43.0.1"
itsdangerous = "^2.2.0"
pyyaml = "^6.0.2"
djangorestframework = "^3.15.2"
django-csp = "^3.8"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-django = "^4.8.0"
ruff = "^0.6.9"
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
